+++ CHANGELOG.md
@@ -25,6 +25,7 @@
 - properly positioning message menu over the top content
 - added support for unknown file type in diff viewer to show without highlighting
 - storing context messages during the agent run
+- added browser notifications
 
 ## [0.50.1]